backend/requirements.txt
tornado==6.3.3
pyyaml==6.0.1
python-dateutil==2.8.2
pytest==7.4.2
pytest-asyncio==0.21.1

# GraphQL dependencies
graphene==3.3
typing-extensions==4.8.0

# DataLoader for batching
aiodataloader==0.4.0
